major refactor
this ports all exported functions to use AnnData objects as input.
Also, reorganization of the code to make it (hopefully) more clear
