+++ Python-module/SpatialDE/_internal/sm_kernel.py
