+++ Python-module/SpatialDE/kernels.py
@@ -5,6 +5,8 @@ import math
 import tensorflow as tf
 from gpflow.utilities.ops import square_distance, difference_matrix
 
+from ._internal.distance_cache import DistanceCache
+
 
 def scale(X: tf.Tensor, lengthscale: Optional[float] = 1):
     if X is not None:
@@ -23,25 +25,6 @@ def scaled_squared_distance(
     X: tf.Tensor, Y: Optional[tf.Tensor] = None, lengthscale: Optional[float] = 1
 ):
     return square_distance(scale(X, lengthscale), scale(Y, lengthscale))
-
-
-class DistanceCache:
-    def __init__(self, X):
-        self.X = X
-        self._squaredEuclidean = None
-        self._sumDiff = None
-
-    @property
-    def squaredEuclideanDistance(self):
-        if self._squaredEuclidean is None:
-            self._squaredEuclidean = square_distance(self.X, None)
-        return self._squaredEuclidean
-
-    @property
-    def sumOfDifferences(self):
-        if self._sumDiff is None:
-            self._sumDiff = tf.reduce_sum(difference_matrix(self.X, None), axis=-1)
-        return self._sumDiff
 
 class Kernel:
     def __init__(self, cache: DistanceCache):
